+++ test/test_asset_report.rb
@@ -2,7 +2,7 @@ require_relative 'test_helper'
 
 # Internal: The test for Plaid::AssetReport.
 class PlaidAssetReportTest < PlaidTest
-  def test_full_flow # rubocop:disable Metrics/AbcSize, Metrics/MethodLength
+  def test_full_flow # rubocop:disable Metrics/MethodLength
     # Create an item with the "assets" product.
     create_item initial_products: [:assets]
 
@@ -84,7 +84,7 @@ end
 
 def poll_for_asset_report(asset_report_token)
   response = nil
-  20.times do
+  50.times do
     begin
       response = @client.asset_report.get(asset_report_token)
       break